requirements.txt
streamlit
pytesseract
Pillow
selenium